requirements.txt
tweepy==4.14.0
kafka-python==2.0.2
pyspark==3.3.0
pymongo==4.5.0
textblob==0.17.1
vaderSentiment==3.3.2
streamlit==1.28.0
plotly==5.17.0
pandas==2.1.1
wordcloud==1.9.2
matplotlib==3.8.0